+++ src/deluge/model/action/action_logger.cpp
@@ -604,6 +604,7 @@ currentClipSwitchedOver:
 	}
 
 	else if (whichAnimation == Animation::EXIT_AUTOMATION_VIEW) {
+		automationView.resetShortcutBlinking();
 		if (getCurrentClip()->type == ClipType::INSTRUMENT) {
 			changeRootUI(&instrumentClipView);
 		}